pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alignrl"
version = "0.4.0"
description = "LLM post-training playbook: SFT, GRPO, DPO, eval, and inference"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Samanvya Tripathi" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
train = [
    "torch>=2.1",
    "transformers>=4.45",
    "trl>=0.12",
    "peft>=0.13",
    "datasets>=3.0",
    "bitsandbytes>=0.44",
    "accelerate>=1.0",
    "wandb",
]
unsloth = ["unsloth"]
eval = ["lm-eval>=0.4"]
serve = ["vllm>=0.6"]
mlx = ["mlx-lm>=0.18"]
gradio = ["gradio>=4.0"]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
    "mypy>=1.10",
]
all = ["alignrl[train,eval,serve,gradio,dev]"]

[project.scripts]
alignrl = "alignrl.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
